requirements.txt
fastapi
pymysql
python-dotenv
pydantic
sqlalchemy
uvicorn